% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conv_byte}
\alias{conv_byte}
\title{function to convert numeric (byte) value into pretty byte representation}
\usage{
conv_byte(size)
}
\arguments{
\item{size}{Numeric size value.}
}
\description{
taken from: https://stackoverflow.com/questions/29787452/how-do-i-quickly-convert-the-size-element-of-file-info-from-bytes-to-kb-mb-g
taken from
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
