% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreport.R
\name{dbreport}
\alias{dbreport}
\title{Create reports of a table}
\usage{
dbreport(
  con = NULL,
  schema = NULL,
  tbl = NULL,
  col = NULL,
  entry = NULL,
  title = "My report title",
  text = NULL,
  plot_distinct = TRUE,
  plot_type = c("lollipop", "treemap"),
  plot_limit = 100L,
  output_dir = NULL,
  output_file = "report",
  output_format = c("html_document", "html_vignette", "pdf_document", "word_document"),
  file = TRUE,
  file_format = c("csv", "json"),
  file_type = c("single", "multiple"),
  exit = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{con}{Connection to data base, R-object (data.table, data.frame, tibble). Required.}

\item{schema}{Data base schema (if applicable).}

\item{tbl}{Table to be used for summary report (if applicable).}

\item{col}{Table columns that should be summarised in report.}

\item{entry}{Table entries which should be comsidered in summary report.}

\item{title}{Title of summary report.}

\item{text}{User defined text input to add some description to the report. Default NULL - no text. Can be a character vector or a file.}

\item{plot_distinct}{Include plots of ordered distinct counts of each column in summary report.}

\item{plot_type}{Should a lollipop or a treemap plot be produced?}

\item{plot_limit}{Limit distinct count axis in the plots. Default: 100L. Should be set to a reasonable size.}

\item{output_dir}{To which directory should the output be saved?}

\item{output_file}{File name}

\item{output_format}{Which markdown output_format should be used for rendering the summary report? E.g. html_document, html_vignette, pdf_document, ...}

\item{file}{Should one or multiple files for each distinct column counts br created?}

\item{file_format}{To which format should the file(s) be put out (csv, json). Multiple possible.}

\item{file_type}{Output each column distinct counts set to a single or to multiple files.}

\item{exit}{logical. Should database connections be closed after runing dbreport(). Default: TRUE.}

\item{verbose}{TODO}
}
\value{
The function creates a markdown report and exports files
}
\description{
Create reports from data base tables, R table objects (data.table, data.frame, tibble)
}
\examples{
\donttest{
# creates a .html, .pdf, .doc via Rmarkdown
dbreport(con = iris,
         plot_distinct = TRUE,
         plot_type = 'lollipop',
         plot_limit = 10, # Limit your plot to the first 15 entries
         output_dir = file.path(tempdir(), 'iris'),
         output_file = 'test',
         output_format = 'pdf_document', # you can also create a .pdf
         title = 'My iris report',
         file = TRUE, # Output single distinct counts to a file
         file_type = 'multiple',
         file_format = 'csv')
}
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
