% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples_n}
\alias{examples_n}
\title{function to transpose example columns}
\usage{
examples_n(l, n = 3)
}
\arguments{
\item{l}{A list.}

\item{n}{Number of examples to be selected.}
}
\description{
function to transpose example columns
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
