% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Function to retrieve meta data from databases or R-objects}
\usage{
meta(...)
}
\arguments{
\item{con}{Database connection or R table object}

\item{schema}{Database schema}

\item{tbl}{Database table}

\item{plot_type}{Plot type}

\item{plot_limit}{Plot limit}

\item{plot_limit_text}{Plot text limit}

\item{output_dir}{Output directory}

\item{output_file}{Output file}
}
\description{
Function to retrieve meta data from databases or R-objects
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
