% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_bar_horiz}
\alias{pl_bar_horiz}
\title{Plots a horizontal bar-ggplot()}
\usage{
pl_bar_horiz(dat = NULL, x = NULL, y = NULL)
}
\arguments{
\item{dat}{data.table object}

\item{x}{category}

\item{y}{count}
}
\description{
Plots a horizontal bar-ggplot()
}
