% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_lollipop}
\alias{pl_lollipop}
\title{Plots a vertical lollipop ggplot()}
\usage{
pl_lollipop(dat = NULL, x = NULL, y = NULL)
}
\arguments{
\item{dat}{data.table object}

\item{x}{category}

\item{y}{count}
}
\description{
Plots a vertical lollipop ggplot()
}
