% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_prepare.R
\name{pl_prepare}
\alias{pl_prepare}
\title{Helper S3 method to prepare data for plots}
\usage{
pl_prepare(...)
}
\arguments{
\item{dat}{data.table object}

\item{col}{column}

\item{plot_limit}{Plot limit}

\item{plot_limit_text}{Plot text limit}
}
\description{
Helper S3 method to prepare data for plots
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
