% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_tree}
\alias{pl_tree}
\title{Plots a treemap()}
\usage{
pl_tree(dat = NULL, index = NULL, vsize = NULL)
}
\arguments{
\item{dat}{data.table object}

\item{index}{used in treemap()}

\item{vsize}{used in treemap()}
}
\description{
Plots a treemap()
}
