% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_treemap}
\alias{pl_treemap}
\title{Plots a ggplot() treemap}
\usage{
pl_treemap(dat = NULL, x = NULL, y = NULL)
}
\arguments{
\item{dat}{data.table object}

\item{x}{category}

\item{y}{count}
}
\description{
Plots a ggplot() treemap
}
