% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_limit}
\alias{str_limit}
\title{Shorten string}
\usage{
str_limit(x, limit = 20L)
}
\arguments{
\item{x}{string}

\item{limit}{integer; number of characters to limit string to}
}
\value{
a string
}
\description{
Shorten string
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
