% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{table_info}
\alias{table_info}
\title{Function to retrieve meta information of many tables.}
\usage{
table_info(con, schema = NULL, tbl = NULL)
}
\arguments{
\item{con}{Database connection or R table object.}

\item{schema}{Database schema.}

\item{tbl}{Database table.}

\item{...}{Currently not used.}
}
\description{
Function to retrieve meta information of many tables.
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
