% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_entry.R
\name{tbl_col_entry}
\alias{tbl_col_entry}
\title{S3 method to get count of specific entries}
\usage{
tbl_col_entry(...)
}
\arguments{
\item{con}{database connection or R table object}

\item{schema}{database schema}

\item{tbl}{database table}

\item{col}{specific table column}

\item{entry}{entry to be looked for}
}
\description{
S3 method to get count of specific entries
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
