% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_type.R
\name{tbl_col_type}
\alias{tbl_col_type}
\title{S3 method to retrieve the column types of a table}
\usage{
tbl_col_type(...)
}
\arguments{
\item{con}{database connection or R table object}

\item{schema}{database schema}

\item{tbl}{database table}
}
\description{
S3 method to retrieve the column types of a table
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
