% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_name.R
\name{tbl_name}
\alias{tbl_name}
\title{S3 method to retrieve name of table}
\usage{
tbl_name(...)
}
\arguments{
\item{con}{database connection or R table object}

\item{schema}{database schema}

\item{tbl}{database table}
}
\description{
S3 method to retrieve name of table
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
