% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_row.R
\name{tbl_row}
\alias{tbl_row}
\title{S3 method to retrieve the number of rows}
\usage{
tbl_row(...)
}
\arguments{
\item{con}{database connection obj or R table object}

\item{schema}{database schema}

\item{tbl}{database table}
}
\description{
S3 method to retrieve the number of rows
}
