% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_size.R
\name{tbl_size}
\alias{tbl_size}
\title{S3 method to retrieve size of table}
\usage{
tbl_size(...)
}
\arguments{
\item{con}{database connection obj or R table object}

\item{schema}{database schema}

\item{tbl}{database table}
}
\description{
S3 method to retrieve size of table
}
