% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verbose_message}
\alias{verbose_message}
\title{Verbose ouput messages}
\usage{
verbose_message(verbose, msg)
}
\arguments{
\item{verbose}{Should a verbose output be printed to console?}

\item{msg}{Message to be printed to console}
}
\description{
Verbose ouput messages
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
