% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_to_dir}
\alias{write_to_dir}
\title{function to write data to a directory}
\usage{
write_to_dir(
  l = NULL,
  output_dir = NULL,
  output_file = NULL,
  file_format = c("csv", "json"),
  file_type = c("single", "multiple")
)
}
\arguments{
\item{output_dir}{Directory to write to.}

\item{output_file}{Name of the output file.}

\item{file_format}{Format of the file.}

\item{file_type}{Should a single or multiple files be written?}
}
\description{
function to write data to a directory
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
