% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreport.R
\name{dbreport}
\alias{dbreport}
\title{Create reports of a table}
\usage{
dbreport(
  con = NULL,
  schema = NULL,
  tbl = NULL,
  col = NULL,
  entry = NULL,
  report_title = "My report title",
  report_text = NULL,
  plot_distinct = TRUE,
  plot_type = c("lollipop", "bar_horiz", "treemap"),
  plot_limit = 40L,
  plot_limit_text = 20L,
  output_dir = NULL,
  output_file = "report",
  output_format = html_document(theme = "united", highlight = "tango"),
  file = TRUE,
  file_format = c("csv", "json"),
  file_type = c("single", "multiple"),
  exit = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{con}{Connection to data base or an R-object (data.table, data.frame, tibble). Required.}

\item{schema}{character; Data base schema (if applicable).}

\item{tbl}{character; Table to be used for summary report (if applicable).}

\item{col}{character; Table columns that should be summarised in report.}

\item{entry}{any; Specific table entries which should be comsidered in the summary table.}

\item{report_title}{character; Title of summary report.}

\item{report_text}{character or connection; User defined text input to add some description to the report. Default NULL - no text. Can be a character vector or a file.}

\item{plot_distinct}{logical; Include plots of ordered distinct counts of each column entryin the report?}

\item{plot_type}{character; Should a lollipop, barplot or treemap plot be produced?}

\item{plot_limit}{integer; Limit distinct count axis in the plots. Default: 40L.}

\item{plot_limit_text}{integer; Limit axis text to a certain length. Default: NULL (i.e. no limit is applied).}

\item{output_dir}{character; To which directory should the output be saved?}

\item{output_file}{character; File name of the report.}

\item{output_format}{object; directly passed on to rmarkdown::render(). Default: html_document(theme = 'united', highlight = 'tango').}

\item{file}{logical; Should one or multiple files for each distinct column counts be created?}

\item{file_format}{character; To which format should the file(s) be put out (csv, json). Multiple possible.}

\item{file_type}{character; Output each column distinct counts set to a single or to multiple files.}

\item{exit}{logical; Should database connections be closed after runing dbreport(). Default: TRUE.}

\item{verbose}{TODO}
}
\value{
The function creates a markdown report and exports files
}
\description{
Create reports from data base tables, R table objects (data.table, data.frame, tibble)
}
\examples{
\donttest{
# creates a .html, .pdf, .doc via Rmarkdown
dbreport(
  con = iris,
  report_title = 'My iris report',
  report_text = 'Here goes my text as an inrtoducion to the report.'
  plot_distinct = TRUE,
  plot_type = 'lollipop',
  plot_limit = 10, # Limit your plot to the first 15 entries
  output_dir = file.path(tempdir(), 'iris'),
  output_file = 'test',
  file = TRUE, # Output single distinct counts to a file
  file_type = 'multiple',
  file_format = 'csv'
)
}
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
