% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Function to retrieve data from databases or R-objects}
\usage{
fetch(
  con = NULL,
  schema = NULL,
  tbl = NULL,
  col = NULL,
  entry = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{con}{Database connection or R table object}

\item{schema}{Database schema}

\item{tbl}{Database table}

\item{col}{Specific table column}

\item{entry}{Entry to be looked for}

\item{verbose}{Verbose output?}
}
\description{
Function to retrieve data from databases or R-objects
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
