% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perc_value}
\alias{perc_value}
\title{function to convert numeric values to perc strings}
\usage{
perc_value(x, total)
}
\arguments{
\item{x}{Numeric vector to convert.}

\item{total}{Numeric vector to divide with.}
}
\description{
function to convert numeric values to perc strings
}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
