% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_bar_horiz}
\alias{pl_bar_horiz}
\title{Plots a horizontal bar-ggplot()}
\usage{
pl_bar_horiz(...)
}
\arguments{
\item{...}{not used}

\item{dat}{data.table object}

\item{column}{Column}

\item{plot_limit}{Limit number of plotted lines}

\item{plot_limit_text}{Limit axis plot axis text}
}
\description{
Plots a horizontal bar-ggplot()
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
