% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_lollipop}
\alias{pl_lollipop}
\title{Plots a vertical lollipop ggplot()}
\usage{
pl_lollipop(...)
}
\arguments{
\item{...}{not used}

\item{dat}{data.table object}

\item{column}{Column}

\item{plot_limit}{Limit number of plotted lines}

\item{plot_limit_text}{Limit axis plot axis text}
}
\description{
Plots a vertical lollipop ggplot()
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
