% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pl_treemap}
\alias{pl_treemap}
\title{Plots a ggplot() treemap}
\usage{
pl_treemap(...)
}
\arguments{
\item{...}{not used}

\item{dat}{data.table object}

\item{col}{Column}

\item{plot_limit}{Limit number of plotted lines}

\item{plot_limit_text}{Limit axis plot axis text}
}
\description{
Plots a ggplot() treemap
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
