% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subchunkify}
\alias{subchunkify}
\title{Create sub-chunks for plots}
\usage{
subchunkify(pl, fig.height = 7, fig.width = 5, chunk_name = "plot")
}
\arguments{
\item{pl}{a plot object}

\item{fig.height}{figure height}

\item{fig.width}{figure width}

\item{chunk_name}{name of the chunk}
}
\description{
Create sub-chunks for plots
}
