% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_distinct.R
\name{tbl_col_distinct}
\alias{tbl_col_distinct}
\title{S3 method to get distinct entries for every column}
\usage{
tbl_col_distinct(...)
}
\arguments{
\item{con}{database connection obj or R table object}

\item{schema}{database schema}

\item{tbl}{database table}

\item{col}{specific table column}

\item{verbose}{verbose output}
}
\description{
S3 method to get distinct entries for every column
}
