% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_distinct.R
\name{tbl_col_distinct}
\alias{tbl_col_distinct}
\title{S3 method to get distinct entries for every column}
\usage{
tbl_col_distinct(...)
}
\arguments{
\item{...}{Currently not used.}

\item{con}{Database connection or R table object.}

\item{schema}{Database schema.}

\item{tbl}{Database table.}

\item{column}{Specific table column.}

\item{col_numeric}{Is a column numeric or not?}

\item{verbose}{Verbose output.}
}
\description{
S3 method to get distinct entries for every column
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
