% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_distinct_n.R
\name{tbl_col_distinct_n}
\alias{tbl_col_distinct_n}
\title{S3 method to get counts of distinct entries for every column}
\usage{
tbl_col_distinct_n(...)
}
\arguments{
\item{con}{database connection or R table object}

\item{schema}{database schema}

\item{tbl}{database table}

\item{column}{specific table column}

\item{verbose}{verbose output}
}
\description{
S3 method to get counts of distinct entries for every column
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
