% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_null.R
\name{tbl_col_null}
\alias{tbl_col_null}
\title{S3 method to get the amount of NULLs or NAs in each column}
\usage{
tbl_col_null(...)
}
\arguments{
\item{con}{database connection obj or R table object}

\item{schema}{database schema}

\item{tbl}{database table}

\item{col}{specific table column}
}
\description{
S3 method to get the amount of NULLs or NAs in each column
}
