% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_to_dir}
\alias{write_to_dir}
\title{function to write data to a directory}
\usage{
write_to_dir(
  l = NULL,
  output_dir = NULL,
  output_file = NULL,
  file_format = c("csv", "json"),
  file_type = c("single", "multiple")
)
}
\arguments{
\item{con}{database connection or R table object}

\item{schema}{database schema}

\item{tbl}{database table}
}
\description{
function to write data to a directory
}
\author{
Andreas Scharmueller, \email{andschar@protonmail.com}
}
